Pin requirements sensibly
This commit ensures that the vcrpy packaged is pinned to the last
major version available for Python 2 surpressing deprecation
warnings. We also make sure it is able to access the later packages
available for Python 3.

Other requirements are pinned to the upper-bound of their major
versions that we've tested with. Or, in the case of Flake8 it has
a lower bound where it is presumed that future updates will only be
more useful to code quality.

Setup.py requirements now match requirements.txt.