Enable file streaming to the command-line and API
Enables streaming to the API and command-line and makes changes to
ensure that other functions can return None to the command line
without stdout being polluted, i.e. output None directly to the CLI.
Tests have been included to make this change as robust as possible
for the consumer. Docstrings have been updated to reflect the
changes.